apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
timm==0.9.16
Pillow==10.3.0
python-multipart==0.0.9